balance-service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
dnspython==2.7.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
kafka-python==2.2.15
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.2
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0